pyproject.toml
[project]
name = "mic2e-inference"
version = "0.1.0"
description = "Inference API for MIC2E"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.35.2",
    "fastapi>=0.123.7",
    "huggingface-hub>=0.36.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scipy>=1.15.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }
